+++ srcs/fra_pix_to_re.c
@@ -6,18 +6,18 @@
 /*   By: erli <erli@42.fr>                          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/07 16:52:15 by erli              #+#    #+#             */
-/*   Updated: 2018/12/12 09:35:48 by erli             ###   ########.fr       */
+/*   Updated: 2018/12/12 11:05:46 by erli             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fractol.h"
 #include "libft.h"
 
-long double		fra_pix_to_re(t_fra_param *param, int px)
+double			fra_pix_to_re(t_fra_param *param, int px)
 {
-	long double real;
+	double	 real;
 
-	real = ((long double)(px * 2) - (long double)IMG_WIDTH) * param->qx
+	real = (double)((px * 2) - IMG_WIDTH) * param->qx
 		+ param->o_re;
 	return (real);
 }
